New boolean to create(or not) the wsdl subfolder based on serviceName

+++ src/main/java/com/github/dkschlos/wsdlget/WsdlGetMojo.java
@@ -1,14 +1,15 @@
 package com.github.dkschlos.wsdlget;
 
 import com.github.dkschlos.wsdlget.internal.WsdlDownloader;
-import java.io.File;
-import java.util.List;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
+
+import java.io.File;
+import java.util.List;
 
 @Mojo(name = "wsdlget", defaultPhase = LifecyclePhase.GENERATE_RESOURCES)
 public class WsdlGetMojo extends AbstractMojo {
@@ -20,16 +21,19 @@ public class WsdlGetMojo extends AbstractMojo {
     private String outputPath;
 
     @Parameter(defaultValue = "false")
+    private boolean subfolderByServiceName;
+
+    @Parameter(defaultValue = "false")
     private boolean clearOutputDirectory;
 
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
         File root = new File(outputPath);
-        if (clearOutputDirectory) {
+        if (clearOutputDirectory && root.exists()) {
             deleteRecursively(root);
         }
         for (WsdlDefinition wsdl : wsdls) {
-            WsdlDownloader downloader = new WsdlDownloader(root, wsdl);
+            WsdlDownloader downloader = new WsdlDownloader(root, wsdl, subfolderByServiceName);
             downloader.download();
         }
     }
